Add custom column type support
When aliasing columns, MongoDB loses type information. This hack allows
manual specification of types for MongoDB. At present it is restricted
to MongoDB, but may be possible to use it for other target DB types if
required. There is no user interface at present as in, no support for
adding this stuff using the command line. That remains to be done.